+++ odbc-api/src/buffers.rs
@@ -1,14 +1,13 @@
 //! This module contains buffers intended to be bound to ODBC statement handles.
-mod fixed_sized;
+mod columnar;
+mod column_with_indicator;
 mod text_column;
 mod text_row_set;
+mod description;
 
 pub use self::{
-    fixed_sized::{
-        OptBitColumn, OptDateColumn, OptF32Column, OptF64Column, OptFixedSizedColumn, OptI32Column,
-        OptI64Column, OptI8Column, OptNumericColumn, OptTimeColumn, OptTimestampColumn,
-        OptU8Column,
-    },
-    text_column::TextColumn,
+    description::{BufferDescription, BufferKind},
+    columnar::{ColumnarRowSet, AnyColumnView},
+    text_column::{TextColumn, TextColumnIt},
     text_row_set::TextRowSet,
 };